Added entities and inheritance
The Ship now inherits from an Entity class. Other entities such as
lasers and landers can do the same in future.
The EntityHolder class acts as a container for all of the enitites in
the game, allowing for the GameLogic to easily keep track of them. This
is used in the moving of all entites and the generating of the Character
list for the UserInterface.
The Ship data member of the GameLogic class has been replaced by a
shared_ptr<Ship> so that the ship can also be contained in the entity
list.