Removed _character from Entity
For more consistent coding style the member data _character has been
removed. This has been replaced with a _position and an _ID data member.
This mean that we can now say _position rather than _character.position,
which is more consistent with our usage of _velocity. Also moved the
facing_right member data from the Character class and moved it to Ship.
This was done because it was a refused bequest for all other entities.